+++ packages/interactive-code-scroll/test/fixtures/broken/tutorial/code/main.js
@@ -1,0 +1,3 @@
+// #region config
+const clientId = "ID"; // @var clientId
+// #endregion